Add an attcacheoff-populating function
It populates attcacheoff-capable attributes with the correct offset,
and fills attributes whose offset is uncacheable with an 'uncacheable'
indicator value; as opposed to -1 which signals "unknown".

This allows users of the API to remove redundant cycles that try to
cache the offset of attributes - instead of O(N-attrs) operations, this
one only requires a O(1) check for repeat calls.