btree: Specialize various performance-sensitive functions on btree key shape.
nbtree keys are not all made the same, and a significant amount of time is
spent in code that exists only to deal with various key shapes. By specializing
function calls based on the key shape, we can remove or reduce these causes
of overhead.

This commit adds the basic infrastructure for specializing specific hot code
in the nbtree AM to certain shapes of keys, and splits the code that can
benefit from attribute offset optimizations into separate files. This does
NOT yet update the code itself - it just makes the code compile cleanly.
The performance should be comparable if not the same.