WIP: Create some "NodeDesc" infrastructure
This is a set of compiled metadata about Nodes, which details information
about how to treat them in a good manner so that we can do metaprogramming
on Node structures at runtime.

The infrastructure is primarily built for reading and writing out node
structures, but will probably also be useful for query jumbling and deep
copying of Node trees.

For now, the infrastructure seems to be somewhat smaller than the existing
implementation; both in binary as in VM size.