NBTree: Reduce Index-Only Scan pinning requirements
Previously, we would keep a pin on every leaf page while we were returning
tuples to the scan.  With this patch, we utilize the newly introduced
table_index_vischeck_tuples API to pre-check visibility of all TIDs, and
thus unpin the page well ahead of when we'd usually be ready with returning
and processing all index tuple results.  This reduces the time VACUUM may
have to wait for a pin, and can increase performance with reduced redundant
VM checks.